// server/src/report/enums/report-template.enum.ts
export enum ReportTemplate {
    PYC_NT = 'PYC_NT',
    BM_Betong = 'BM_Betong',
    BM_Mong = 'BM_Mong',
    BM_Tru = 'BM_Tru',
    BM_Neo = 'BM_Neo',
    BM_KeoDay = 'BM_KeoDay',
    BM_TiepDia = 'BM_TiepDia',
    BM_Tram = 'BM_Tram',
    BM_LapDatPD = 'BM_LapDatPD',
    BM_ThuHoi = 'BM_ThuHoi',
}